+++ llama_index/retrievers/__init__.py
@@ -1,7 +1,10 @@
 from llama_index.indices.base_retriever import BaseRetriever
 from llama_index.indices.empty.retrievers import EmptyIndexRetriever
 from llama_index.indices.keyword_table.retrievers import KeywordTableSimpleRetriever
-from llama_index.indices.knowledge_graph.retriever import KGTableRetriever
+from llama_index.indices.knowledge_graph.retrievers import (
+    KGTableRetriever,
+    KnowledgeGraphRAGRetriever,
+)
 from llama_index.indices.list.retrievers import (
     ListIndexEmbeddingRetriever,
     ListIndexRetriever,
@@ -26,6 +29,7 @@ __all__ = [
     "ListIndexEmbeddingRetriever",
     "ListIndexRetriever",
     "KGTableRetriever",
+    "KnowledgeGraphRAGRetriever",
     "EmptyIndexRetriever",
     "TreeAllLeafRetriever",
     "TreeSelectLeafEmbeddingRetriever",